fetch data from upower dbus and display on screen reactively

+++ src/models/battery_data.rs
@@ -1,11 +1,11 @@
 #[derive(Clone, PartialEq, Debug)]
 pub struct BatteryData {
-    pub percentage: f32,
+    pub percentage: f64,
     pub charging: bool,
-    pub health: f32,
-    pub temperature: f32,
-    pub rate: f32,
-    pub time_remaining: f32,
+    pub health: f64,
+    pub temperature: f64,
+    pub rate: f64,
+    pub time_remaining: i64,
     pub history_percentage: Vec<(f64, f32)>,
     pub history_rate: Vec<(f64, f32)>,
 }
